DEV: Convert connector component to Glimmer component
Why this change?

We think the declaring a proper Ember component directly and then using
the component in plugin outlet connector is much clearer than relying on
the `setupComponent` hook because authors will just be using Ember
component patterns which are well documented instead of some custom
pattern that is defined by "Discourse connectors".