DEV: Remove default value from `sections` (#50)
This commit removes the default values from the `sections` theme setting
since they are too generic and are not good defaults. The defaults were
useful previously for the old `link_sections` and `links` theme settings
since there was an "implicit" format to follow and the defaults serve as
examples of the format. Now that `sections` is a objects type setting,
there is no need for us to provide any examples.